fixed test, snapshot, and method stub for the topAgentsOfDeath method (#7)
The test written for the final challenge in this exercise was erroneously calling the second to last method (elevatedVolcanoes) instead of the actual last method to implement (topAgentsOfDeath). Because of this, the snapshot initially generated by Jest was incorrect, as well. I've updated the test, snapshot, and source file for volcano.js to fix this issue. Check it out!